mk more dir & change
Recently, I got too many things to do. Thats why i'm a little tired. When i turn back to check my file and progress, i found that i don't even know a file in which dir . so it's a lot of change.  like Dr,song said, i shoule make everthing in one system, so that i can know what it is in the first sight.